backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
requests==2.31.0
python-dotenv==1.0.0
langchain==0.1.4
# openai==1.12.0
# anthropic==0.8.1
pydantic==2.5.3